Addressed review comments-3
